+++ lib/tldr/rake.rb
@@ -16,13 +16,17 @@ class TLDR
     def define name, task_config
       desc "Run #{name} tests (use TLDR_OPTS or .tldr.yml to configure)"
       task name do
-        cli_args = build_cli_args(task_config)
-        fail unless system "#{"bundle exec " if defined?(Bundler)}tldr #{cli_args}"
+        argv = Shellwords.shellwords(merge_env_opts(task_config).to_full_args)
+        begin
+          TLDR::Run.cli(argv)
+        rescue SystemExit => e
+          fail "TLDR task #{name} failed with status #{e.status}" unless e.status == 0
+        end
       end
     end
 
-    def build_cli_args task_config
-      config = if ENV["TLDR_OPTS"]
+    def merge_env_opts task_config
+      if ENV["TLDR_OPTS"]
         env_argv = Shellwords.shellwords(ENV["TLDR_OPTS"])
         opts_config = ArgvParser.new.parse(env_argv, {
           config_intended_for_merge_only: true
@@ -31,8 +35,6 @@ class TLDR
       else
         task_config
       end
-
-      config.to_full_args
     end
   end
 end