+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Stop shelling out to `tldr` from our Rake task. Rescue `SystemExit` instead
+
 ## [0.5.0]
 
 * Define your own Rake tasks with `TLDR::Task` and pass in a custom configuration